Move normalizeRadians() to new file MathUtils. Add test.
