+++ src/MathUtils.ts
@@ -1,0 +1,11 @@
+// Copyright 2019 Cognite AS
+
+export function normalizeRadians(angle: number, lowerBound = -Math.PI, upperBound = Math.PI): number {
+  while (angle < lowerBound) {
+    angle += 2 * Math.PI;
+  }
+  while (angle > upperBound) {
+    angle -= 2 * Math.PI;
+  }
+  return angle;
+}
